// frontend/types/story.ts
export enum CreateStoryView {
    SELECTION = "SELECTION",
    JOURNEY = "JOURNEY",
    BEFORE_AFTER = "BEFORE_AFTER",
    CHALLENGE = "CHALLENGE",
    ALBUM = "ALBUM",
}

export enum LeaveStoryAction {
    KEEP_AS_INDEPENDENT_MOMENTS = "KEEP_AS_INDEPENDENT_MOMENTS",
    REMOVE_MOMENTS = "REMOVE_MOMENTS",
    CREATE_PERSONAL_STORY = "CREATE_PERSONAL_STORY",
}